requirements.txt
django_debug_toolbar==3.2.2
reportlab==3.6.1
xlwt==1.3.0
django_redis==5.0.0
requests==2.25.1
Django==3.2.7
bpmappers==1.2
djangorestframework==3.12.4
Pillow==8.3.2
PyJWT==2.1.0
qiniu==7.4.1
